Restructure app.ts to different sub-moduls.
