+++ src/models/user.ts
@@ -1,7 +1,4 @@
-//For in-memory storage:
-let users: User[] = [];
-
-class User {
+export class User {
   userId: string;
   displayName: string;
   mail: string;
@@ -13,17 +10,5 @@ class User {
     this.mail = mail;
     this.userPrincipalName = userPrincipalName;
   }
-
-  static findOrCreate(profile: any, done: any) {
-    let user = users.find(u => u.userId === profile.id);
-    if (user) {
-      done(null, user);
-    } else {
-      user = new User(profile.id, profile.displayName, profile.mail, profile.userPrincipalName);
-      users.push(user);
-      done(null, user);
-    }
-  }
 }
 
-export default User;
