+++ src/structure.txt
@@ -1,0 +1,17 @@
+db/
+dist/
+  models/
+    user.js
+    userStorage.js
+  app.js
+node_modules/
+public/
+src/
+  models/
+    user.ts
+    userStorage.ts
+  routes/
+    authRoutes.ts
+    mainRoutes.ts
+    logoutRoute.ts
+  app.ts
